Koniec pod obronę

+++ Project/invoices/views.py
@@ -2,6 +2,7 @@ from django.shortcuts import render
 from basket.models import Basket
 from .models import Invoice, Invoices
 from django.contrib.auth.decorators import login_required
+from .filters import SnippetFilter
 
 
 @login_required(login_url='/login')
@@ -10,17 +11,21 @@ def show_invoices(request):
     logged_in = False
     if request.user.is_authenticated:
         logged_in = True
-    inv_base = Invoice.objects.all()
     invs_base = Invoices.objects.all()
     n_objects = Basket.objects.all().count()
     if request.method == "POST":
-        instance = Invoice.objects.filter(
-            invoice_number=request.POST['form_invoice_number'])
-        instance.delete()
+        if 'form_invoice_number' in request.POST:
+            instance = Invoice.objects.filter(
+                invoice_number=request.POST['form_invoice_number'])
+            instance.delete()
+        elif 'show_details' in request.POST:
+            invs_base = Invoices.objects.filter(invoice = request.POST["show_details"])
+        else:
+            pass
     context = {
         'logged_in': logged_in,
-        'inv_base': inv_base,
         'invs_base': invs_base,
         'n_objects': n_objects,
+        'filter': SnippetFilter(request.GET),
     }
     return render(request, "invoices.html", context)
